requirements.txt
# Frontend
streamlit==1.22.0
streamlit-chat==0.0.2.2

# Backend
flask==2.2.3
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
requests==2.28.2

# Data storage
pymongo==4.3.3
motor==3.1.2

# Data processing and ML
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
scipy==1.10.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0 